# Open addressing algorithm for RubyHash