… core/src/main/java/org/jruby/RubyNil.java
… core/src/main/java/org/jruby/ext/pathname/RubyPathname.java
… core/src/main/java/org/jruby/ir/runtime/IRRuntimeHelpers.java
… core/src/main/java/org/jruby/java/proxies/MapJavaProxy.java
… core/src/main/java/org/jruby/runtime/Helpers.java