@headius 2018-06-12
I modified your benchmark to use benchmark-ips and simple loops, since plain benchmark tends not to exercise methods well and block-based loops have much more overhead than simple loops.

```ruby
require 'benchmark/ips'
N = 50_000
small = 1000
tmp = {}
Benchmark.ips do |x|
  x.report("Add:") { i = 0; while i < N; i+=1; tmp[small] = true; end }
  x.report("Fetch:") { i = 0; while i < N; i+=1; tmp[small]; end }
end
```

Here's numbers with standard options, invokedynamic, and invokedynamic on GraalVM.

Before your work:
```
[] ~/projects/jruby $ jruby hash_bench.rb 
Warming up --------------------------------------
                Add:    48.000  i/100ms
              Fetch:    62.000  i/100ms
Calculating -------------------------------------
                Add:    589.298  (± 5.1%) i/s -      2.976k in   5.065141s
              Fetch:    677.965  (± 5.8%) i/s -      3.410k in   5.049201s

[] ~/projects/jruby $ jruby -Xcompile.invokedynamic hash_bench.rb 
Warming up --------------------------------------
                Add:    70.000  i/100ms
              Fetch:    95.000  i/100ms
Calculating -------------------------------------
                Add:    912.031  (± 8.2%) i/s -      4.550k in   5.029019s
              Fetch:      1.025k (± 5.1%) i/s -      5.130k in   5.020724s

[] ~/projects/jruby $ (pickjdk 1 ; jruby -Xcompile.invokedynamic -Xfixnum.cache=false hash_bench.rb)
New JDK: graalvm
Warming up --------------------------------------
                Add:    97.000  i/100ms
              Fetch:   153.000  i/100ms
Calculating -------------------------------------
                Add:      1.487k (± 2.7%) i/s -      7.469k in   5.027800s
              Fetch:      1.744k (± 5.3%) i/s -      8.721k in   5.022021s
```

And after your work:
```
[] ~/projects/jruby $ jruby hash_bench.rb 
Warming up --------------------------------------
                Add:    59.000  i/100ms
              Fetch:    65.000  i/100ms
Calculating -------------------------------------
                Add:    723.540  (± 6.8%) i/s -      3.599k in   5.002143s
              Fetch:    726.660  (± 5.4%) i/s -      3.640k in   5.026187s

[] ~/projects/jruby $ jruby -Xcompile.invokedynamic hash_bench.rb 
Warming up --------------------------------------
                Add:    73.000  i/100ms
              Fetch:    98.000  i/100ms
Calculating -------------------------------------
                Add:    935.883  (±12.6%) i/s -      4.599k in   5.041101s
              Fetch:      1.190k (± 8.2%) i/s -      5.978k in   5.071986s

[] ~/projects/jruby $ (pickjdk 1 ; jruby -Xcompile.invokedynamic -Xfixnum.cache=false hash_bench.rb)
New JDK: graalvm
Warming up --------------------------------------
                Add:   116.000  i/100ms
              Fetch:   173.000  i/100ms
Calculating -------------------------------------
                Add:      1.634k (± 3.8%) i/s -      8.236k in   5.048905s
              Fetch:      1.959k (± 4.3%) i/s -      9.861k in   5.048234s
```

Looks like a good win so far!