@headius 2018-06-12
I did a size comparison. For a Hash with 100_000 entries mapping 0-99_000 to 1, the old impl is 11.847MB in memory and the new version is 11.382MB. If we subtract some of the common objects:

* RubyHashEntry: 37 bytes * 100_000 = 3.7MB
* RubyFixnum: 56 bytes * 100_000 = 5.6MB
* Total shared between old and new (roughly): 9.3

That gives us old and new sizes of 2.547 and 2.082, a savings of 0.465MB or over 18% on the base RubyHash structure. That's pretty good!

So based on that I thought I'd try some small hashes like `{a:1}`, which we obviously see a lot of in Ruby too.

Subtracting the size of the symbol and fixnum, this hash is 166 bytes on the new impl and 52 bytes on the old impl, so this may be justification for packing simple hashes down a bit tighter.