@ChrisBr 2018-06-12
@headius thanks for the benchmark :+1: 

> Subtracting the size of the symbol and fixnum, this hash is 166 bytes on the new impl and 52 bytes on the old impl, so this may be justification for packing simple hashes down a bit tighter.

Ok! I already had some code for a LinearSearch approach without the bins array. I will try to merge these two implementations tonight.

I will also look into getting rid of the RubyHashEntry elements completely, discussed yesterday with @lopex and extracting it into a Util class.

Stay tuned :guitar: