@headius 2018-06-12
If we can get rid of the RubyHashEntry objects that would be a *tremendous* memory savings.