@ChrisBr 2018-06-12
> If we can get rid of the RubyHashEntry objects that would be a tremendous memory savings.

I pushed an initial version of getting rid of the RubyHashEntries. For now, I don't store them anymore but use them as temporary objects. We now store the key and values in separate rows in the entries array. Key's are stored in the even rows and values in the odd rows. This adds quite some complexity from my POV :cry: 

I haven't benchmarked it yet as I think there is a bug in the script you posted (you never increase small variable, do you?). Would be great if you could have a look at it again and run your benchmarks @headius.

I also tried to move this to a Util class but I have some open questions:
- I would need to access the flags in the Utils class (e.g. COMPARE_BY_IDENTITY_F). How would I do that?
- We also would need to implement the iterator, checkIterating, checkModify etc. I guess this would be quite some work ...

However, I we want to implement the LinearSearchHash for small hashes, I think it would be very beneficial if we could extract these two classes...