@ChrisBr 2018-06-13
@headius I added now a linear search approach for small hashes (< 16 elements).