@ChrisBr 2018-06-19
@headius any update regarding your benchmarking script or how I should proceed?