@enebo 2018-06-19
@ChrisBr he is on vacation this week.  You could fix the bench and generate some numbers and see if removing the extra types helped (which of course means running fixed benchmark with original RubyHashEntry - linear search.  Memory we can probably calculate manually.  We would largely remove 3.7M - field size.