@ChrisBr 2018-06-20
@enebo okay! Here you go:

This is the script I used:
```
require 'benchmark/ips'

[3, 15, 16, 100, 1_000, 10_000].each do |n|
  puts "=========== #{n} elements ============="
  tmp = {}
  Benchmark.ips do |x|
    x.report("Add:") { i = 0; while i < n; tmp[i] = true; i += 1; end }
    x.report("Fetch:") { i = 0; while i < n; tmp[i]; i += 1; end }
  end
end
```

### Benchmark of this PR
* linear search for hash size < 16
* open addressing for hash size > 16
* "no" HashEntry elements

```
=========== 3 elements =============
Warming up --------------------------------------
                Add:    97.614k i/100ms
              Fetch:   103.702k i/100ms
Calculating -------------------------------------
                Add:      3.957M (± 6.4%) i/s -     19.718M in   5.005939s
              Fetch:      3.560M (± 7.3%) i/s -     17.733M in   5.009620s

=========== 15 elements =============
Warming up --------------------------------------
                Add:    39.879k i/100ms
              Fetch:    24.936k i/100ms
Calculating -------------------------------------
                Add:    306.576k (± 4.2%) i/s -      1.555M in   5.082280s
              Fetch:    316.506k (± 4.4%) i/s -      1.596M in   5.052030s

=========== 16 elements =============
Warming up --------------------------------------
                Add:    63.316k i/100ms
              Fetch:    67.519k i/100ms
Calculating -------------------------------------
                Add:      1.243M (± 2.8%) i/s -      6.268M in   5.048272s
              Fetch:      1.438M (± 3.1%) i/s -      7.224M in   5.029939s

=========== 100 elements =============
Warming up --------------------------------------
                Add:    19.133k i/100ms
              Fetch:    20.735k i/100ms
Calculating -------------------------------------
                Add:    233.558k (± 1.4%) i/s -      1.186M in   5.079982s
              Fetch:    259.941k (± 3.8%) i/s -      1.306M in   5.033026s

=========== 1_000 elements =============
Warming up --------------------------------------
                Add:     1.846k i/100ms
              Fetch:     1.880k i/100ms
Calculating -------------------------------------
                Add:     19.931k (± 9.9%) i/s -     99.684k in   5.058294s
              Fetch:     21.963k (± 4.1%) i/s -    110.920k in   5.059002s

=========== 10_000 elements =============
Warming up --------------------------------------
                Add:   182.000  i/100ms
              Fetch:   201.000  i/100ms
Calculating -------------------------------------
                Add:      1.960k (± 3.9%) i/s -      9.828k in   5.021143s
              Fetch:      2.060k (± 3.0%) i/s -     10.452k in   5.077663s
```

### Benchmark open addressing with HashEntry elements, no linear search
```
=========== 3 elements =============
Warming up --------------------------------------
                Add:    88.846k i/100ms
              Fetch:    89.483k i/100ms
Calculating -------------------------------------
                Add:      3.946M (± 5.4%) i/s -     19.724M in   5.015148s
              Fetch:      4.206M (± 5.1%) i/s -     21.028M in   5.015265s
=========== 15 elements =============
Warming up --------------------------------------
                Add:    61.520k i/100ms
              Fetch:    59.891k i/100ms
Calculating -------------------------------------
                Add:      1.186M (± 3.2%) i/s -      5.967M in   5.034577s
              Fetch:      1.113M (± 8.4%) i/s -      5.570M in   5.042245s

=========== 16 elements =============
Warming up --------------------------------------
                Add:    54.942k i/100ms
              Fetch:    60.876k i/100ms
Calculating -------------------------------------
                Add:      1.060M (± 6.6%) i/s -      5.329M in   5.050794s
              Fetch:      1.119M (± 4.2%) i/s -      5.600M in   5.012535s

=========== 100 elements =============
Warming up --------------------------------------
                Add:    16.894k i/100ms
              Fetch:    17.104k i/100ms
Calculating -------------------------------------
                Add:    197.118k (± 5.6%) i/s -    996.746k in   5.073804s
              Fetch:    184.110k (± 6.7%) i/s -    923.616k in   5.038682s

=========== 1_000 elements =============
Warming up --------------------------------------
                Add:     2.071k i/100ms
              Fetch:     2.644k i/100ms
Calculating -------------------------------------
                Add:     19.272k (± 3.6%) i/s -     97.337k in   5.057518s
              Fetch:     28.354k (± 9.2%) i/s -    142.776k in   5.081599s

=========== 10_000 elements =============
Warming up --------------------------------------
                Add:   186.000  i/100ms
              Fetch:   254.000  i/100ms
Calculating -------------------------------------
                Add:      1.926k (± 4.2%) i/s -      9.672k in   5.031638s
              Fetch:      3.073k (± 5.0%) i/s -     15.494k in   5.054577s
```

### Benchmark master

```
=========== 3 elements =============
Warming up --------------------------------------
                Add:    84.669k i/100ms
              Fetch:    90.737k i/100ms
Calculating -------------------------------------
                Add:      3.554M (± 7.2%) i/s -     17.696M in   5.007616s
              Fetch:      3.981M (± 5.9%) i/s -     19.871M in   5.009714s

=========== 15 elements =============
Warming up --------------------------------------
                Add:    59.260k i/100ms
              Fetch:    60.791k i/100ms
Calculating -------------------------------------
                Add:      1.034M (± 6.7%) i/s -      5.156M in   5.011235s
              Fetch:      1.221M (± 4.2%) i/s -      6.140M in   5.038107s

=========== 16 elements =============
Warming up --------------------------------------
                Add:    57.091k i/100ms
              Fetch:    58.878k i/100ms
Calculating -------------------------------------
                Add:      1.063M (± 4.2%) i/s -      5.309M in   5.001955s
              Fetch:      1.148M (± 3.9%) i/s -      5.770M in   5.036016s

=========== 100 elements =============
Warming up --------------------------------------
                Add:    13.132k i/100ms
              Fetch:    15.879k i/100ms
Calculating -------------------------------------
                Add:    171.201k (± 5.3%) i/s -    866.712k in   5.077764s
              Fetch:    190.280k (± 3.4%) i/s -    952.740k in   5.013155s

=========== 1_000 elements=============
Warming up --------------------------------------
                Add:     2.091k i/100ms
              Fetch:     2.384k i/100ms
Calculating -------------------------------------
                Add:     18.318k (± 4.8%) i/s -     92.004k in   5.034774s
              Fetch:     24.077k (± 8.3%) i/s -    121.584k in   5.085813s

=========== 10_000 elements =============
Warming up --------------------------------------
                Add:   159.000  i/100ms
              Fetch:   207.000  i/100ms
Calculating -------------------------------------
                Add:      1.651k (± 4.4%) i/s -      8.268k in   5.018704s
              Fetch:      2.065k (± 6.1%) i/s -     10.350k in   5.030614s
```

### Summary
Looks like linear search approach for 15 elements (the max size for using linear search in MRI) is quite slower than hashing (open/closed addressing). We could try to use a smaller array (e.g. only using for 8 elements) with the cost of adding one more resize.

Also 1_000 / 10_000 elements seems to be slower without HashEntry objects. Maybe because the array is general bigger?
For 1_000 elements closed addressing (master) seems even to be a little faster, no idea why this could be?

I don't have a version for linear search with HashEntry objects, @enebo do you think it makes sense to implement such a version for comparison?

Thanks